# trailmemo/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_trailmemo_pkg_response.Response:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      error_code:
        example: ROUTE_NOT_FOUND
        type: string
      message:
        example: success
        type: string
      request_id:
        type: string
    type: object
...
  internal_handler.ToggleLikeRequest:
    properties:
      target_id:
        example: 1
        type: integer
      target_type:
        example: post
        type: string
    required:
    - target_id
    - target_type
    type: object
...
host: localhost:8087
info:
  contact:
    email: support@example.com
    name: API Support
  description: 迹忆旅图寻迹旅游后端 API 文档
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://example.com/terms
  title: TrailMemo API
  version: 1.0.0
paths:
...
  /api/v1/likes/count:
    get:
      consumes:
      - application/json
      description: 获取目标内容的点赞数量
      parameters:
      - description: 目标ID
        in: query
        name: target_id
        required: true
        type: integer
      - description: 目标类型
        in: query
        name: target_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取点赞数量
      tags:
      - 点赞
  /api/v1/likes/status:
    get:
      consumes:
      - application/json
      description: 检查用户是否已对目标进行点赞
      parameters:
      - description: 目标ID
        in: query
        name: target_id
        required: true
        type: integer
      - description: 目标类型
        in: query
        name: target_type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取点赞状态
      tags:
      - 点赞
  /api/v1/likes/toggle:
    post:
      consumes:
      - application/json
      description: 对分享或评论进行点赞或取消点赞
      parameters:
      - description: 点赞信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.ToggleLikeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 点赞/取消点赞
      tags:
      - 点赞
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"